+++ cfgov/cfgov/settings/base.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'localflavor',
     'v1',
     'core',
     'sheerlike',
@@ -120,6 +121,9 @@ STATIC_URL = '/static/'
 
 # Absolute path to the directory static files should be collected to.
 STATIC_ROOT = os.environ.get('DJANGO_STATIC_ROOT', '/var/www/html/static')
+
+MEDIA_ROOT = os.path.join(PROJECT_ROOT, 'f')
+MEDIA_URL = '/f/'
 
 # List of finder classes that know how to find static files in
 # various locations.