+++ cfgov/v1/models.py
@@ -1,17 +1,20 @@
 from django.db import models
 from django.http import Http404
 from django.conf import settings
-
-from wagtail.wagtailcore.models import Page, PagePermissionTester, UserPagePermissionsProxy
-from wagtail.wagtailcore.url_routing import RouteResult
-from wagtail.wagtailcore.fields import RichTextField
-from wagtail.wagtailadmin.edit_handlers import FieldPanel, MultiFieldPanel
-from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
-
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor.us.models import USStateField
 
-class V1Page(Page):
+from wagtail.wagtailcore.models import Page, PagePermissionTester, UserPagePermissionsProxy
+from wagtail.wagtailcore.fields import RichTextField, StreamField
+from wagtail.wagtailadmin.edit_handlers import FieldPanel, MultiFieldPanel, FieldRowPanel, TabbedInterface, ObjectList, StreamFieldPanel
+from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
+from wagtail.wagtaildocs.edit_handlers import DocumentChooserPanel
+from wagtail.wagtailcore import blocks
+from wagtail.wagtailcore.url_routing import RouteResult
+
+
+class CFGOVPage(Page):
     shared = models.BooleanField(default=False)
 
     # This is used solely for subclassing pages we want to make at the CFPB.
@@ -77,13 +80,13 @@ class V1Page(Page):
 
     def permissions_for_user(self, user):
         """
-        Return a V1PagePermissionsTester object defining what actions the user can perform on this page
+        Return a CFGOVPagePermissionTester object defining what actions the user can perform on this page
         """
-        user_perms = V1UserPagePermissionsProxy(user)
+        user_perms = CFGOVUserPagePermissionsProxy(user)
         return user_perms.for_page(self)
 
 
-class V1PagePermissionTester(PagePermissionTester):
+class CFGOVPagePermissionTester(PagePermissionTester):
     def can_unshare(self):
         if not self.user.is_active:
             return False
@@ -103,31 +106,126 @@ class V1PagePermissionTester(PagePermissionTester):
         return self.user.is_superuser or ('edit' in self.permissions)
 
 
-class V1UserPagePermissionsProxy(UserPagePermissionsProxy):
+class CFGOVUserPagePermissionsProxy(UserPagePermissionsProxy):
     def for_page(self, page):
-        """Return a V1PagePermissionTester object that can be used to query
+        """Return a CFGOVPagePermissionTester object that can be used to query
             whether this user has permission to perform specific tasks on the
             given page."""
-        return V1PagePermissionTester(self, page)
+        return CFGOVPagePermissionTester(self, page)
 
 
-class BlogPage(V1Page):
-    body = RichTextField()
-    date = models.DateField("Post date")
-    feed_image = models.ForeignKey(
+class AgendaItemBlock(blocks.StructBlock):
+    start_dt = blocks.DateTimeBlock(required=False, format='%Y-%m-%d %H:%M')
+    end_dt = blocks.DateTimeBlock(required=False, format='%Y-%m-%d %H:%M')
+    description = blocks.CharBlock(max_length=100, required=False)
+    location = blocks.CharBlock(max_length=100, required=False)
+    speakers = blocks.ListBlock(blocks.StructBlock([
+        ('name', blocks.CharBlock(required=False)),
+        ('url', blocks.URLBlock(required=False)),
+        ], icon='user', required=False)
+    )
+
+    class Meta:
+        icon = 'date'
+
+
+class EventPage(CFGOVPage):
+    # General content fields
+    body = RichTextField(blank=True)
+    archive_body = RichTextField(blank=True)
+    live_body = RichTextField(blank=True)
+    future_body = RichTextField(blank=True)
+    start_dt = models.DateField("Starts", blank=True, null=True)
+    end_dt = models.DateField("Ends", blank=True, null=True)
+    future_body = RichTextField(blank=True)
+    archive_image = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
         blank=True,
         on_delete=models.SET_NULL,
         related_name='+'
     )
+    video_transcript = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+    speech_transcript = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+    flickr_url = models.URLField("Flikr URL", blank=True)
+    youtube_url = models.URLField("Youtube URL", blank=True)
+    live_stream_availability = models.BooleanField("Streaming?", default=False, blank=True)
+    live_stream_url = models.URLField("URL", blank=True)
+    live_stream_date = models.DateField("Go Live Date", blank=True, null=True)
+    # Venue content fields
+    venue_name = models.CharField(max_length=100, blank=True)
+    venue_street = models.CharField(max_length=100, blank=True)
+    venue_suite = models.CharField(max_length=100, blank=True)
+    venue_city = models.CharField(max_length=100, blank=True)
+    venue_state = USStateField(blank=True)
+    venue_zip = models.IntegerField(blank=True, null=True)
+    agenda_items = StreamField([('item', AgendaItemBlock())], blank=True)
 
-    content_panels = Page.content_panels + [
-        FieldPanel('date'),
+    # General content tab
+    content_panels = CFGOVPage.content_panels + [
         FieldPanel('body', classname="full"),
+        FieldRowPanel([
+            FieldPanel('start_dt', classname="col6"),
+            FieldPanel('end_dt', classname="col6"),
+        ]),
+        MultiFieldPanel([
+            FieldPanel('archive_body', classname="full"),
+            ImageChooserPanel('archive_image'),
+            DocumentChooserPanel('video_transcript'),
+            DocumentChooserPanel('speech_transcript'),
+            FieldPanel('flickr_url'),
+            FieldPanel('youtube_url'),
+        ], heading='Archive Information'),
+        FieldPanel('live_body', classname="full"),
+        FieldPanel('future_body', classname="full"),
+        MultiFieldPanel([
+            FieldPanel('live_stream_availability'),
+            FieldPanel('live_stream_url'),
+            FieldPanel('live_stream_date'),
+        ], heading='Live Stream Information'),
     ]
-
+    # Venue content tab
+    venue_panels = [
+        FieldPanel('venue_name'),
+        MultiFieldPanel([
+            FieldPanel('venue_street'),
+            FieldPanel('venue_suite'),
+            FieldPanel('venue_city'),
+            FieldPanel('venue_state'),
+            FieldPanel('venue_zip'),
+        ], heading='Venue Address'),
+    ]
+    # Agenda content tab
+    agenda_panels = [
+        StreamFieldPanel('agenda_items'),
+    ]
+    # Promotion panels
     promote_panels = [
-        MultiFieldPanel(Page.promote_panels, "Common page configuration"),
-        ImageChooserPanel('feed_image'),
+        MultiFieldPanel(CFGOVPage.promote_panels, "Common page configuration"),
     ]
+    # Tab handler interface
+    edit_handler = TabbedInterface([
+        ObjectList(content_panels, heading='General Content'),
+        ObjectList(venue_panels, heading='Venue Information'),
+        ObjectList(agenda_panels, heading='Agenda Information'),
+        ObjectList(CFGOVPage.promote_panels, heading='Promote'),
+        ObjectList(CFGOVPage.settings_panels, heading='Settings', classname="settings"),
+    ])
+
+    parent_page_types = ['v1.EventLandingPage']
+
+
+class EventLandingPage(CFGOVPage):
+    subpage_types = ['EventPage']